[prpc] Document how to disable request/response size limits.
Make sure no overflows happen when using maximum possible limits.

An edge case is 32-bit platforms where math.MaxInt is actually
just 2GB and responses of this size are within the realm of
possibility.

R=iannucci@chromium.org

Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5976663
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>

+++ grpc/prpc/encoding.go
@@ -137,7 +137,7 @@ func writeResponse(ctx context.Context, w http.ResponseWriter, res *response) {
 	}
 
 	// Skip sending the response that exceeds the client's limit.
-	if res.maxResponseSize > 0 && len(body) > int(res.maxResponseSize) {
+	if res.maxResponseSize > 0 && int64(len(body)) > res.maxResponseSize {
 		writeError(ctx, w, errResponseTooBig(int64(len(body)), res.maxResponseSize), res.fmt)
 		return
 	}
